Some Major code refractoring related to Database
Organised JDBC as issues arose based on position in project and not
being acknowledged as a library.
DatabaseConnection class had "Major" code refractor to allow for a
stable database connection, as before it was returning null no matter
what.
Added a new Class called DetectUser which will detect what kind of user
is currently logged in.
Created a new package called reference that will contain text files for
different types of references such as code.
TODO test views on mainWindow.