Major Update
Near finalised CourseView and UpdateCourseView, minor changes may come.
Convert most variables in the above to final.
Modified the DatabaseConnection class in relation to the TimetableLogin
class by creating a constructor for the connection and moving the method
call to the constructor, then initializing the Connection in Login.
Commented out currently unused code causing major and minor errors for
better flow.
Convert certain variables in the MainWindow class to final.
Commented out most of the inner class MainView as currently unused code
that once the views have been finalised will be returned to and looked
at for better flow.
Converted variables in TimetableLogin class to final and changed how the
connection worked.
Added javadoc's to the project, which is to be updated regularly.