Code Cleaning, Code Refractoring, Database Structural Changes
Removed any unused classes, variables, and methods.
Major refractoring on most of the classes, changed code signature types,
names, and some values.
- Changed the access modifiers on all variables found
- Set any variables found that could be final to final
Changed the database structure
- modified field types
- changed field names
- rearranged the relationships
- created a user object table
- included static data in departments, rooms, and schools
- included sample users
Created the controller classes for database modelling to GUI.
Added some additional functionality to login and main, though minor at
that.
- A check whether users exist in the system
- logout button on main returns to login

###ERROR###
When you hit the login button on TimetableLogin it exits the system
despite no code telling it to do so, in fact the code tells it to launch
the main window