Rearranged Classes and Variables, Mapped Database through controllers
Modified the structure of the project folder switching around the
classes in the com.project.controller package with the classes in the
com.project.database package.
Updated and modified the User classes along with the other controller
classes to have the constructor match the sql format coming.
Updated the login to check if the user already exists in the system.
-- need to fix it so that everyone when logging in can check, currently
only the 'timetable' login can check.
Added the initial structure for changing the views dynamically based on
who's logged in.
Added more todo's.
Completed a few todo's